pyproject.toml
[tool.poetry]
name = "data_stone"
version = "0.1.0"
description = ""
authors = ["isaque <isaque_dev@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
psycopg2 = "^2.9.1"
gunicorn = "^20.1.0"
django-admin = "^2.0.1"
python-decouple = "^3.4"
mkdocs = "^1.2.2"
forex-python = "^1.6"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
ipython = "^7.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
